+++ game/main.py
@@ -36,15 +36,15 @@ def show_player(self):
     #     return word, coords, orientation
 
 
-def main():
-        player_count = get_player_count()
-        game = ScrabbleGame(player_count)
-        while game.is_playing():
-            show_board(game.get_board())
-            show_player(*game.get_current_player())
-            word, coords, orientation =get_inputs()
-            try:
-                game.play(word, coords, orientation)
-            except Exception as e:
-                 print(e)
+# def main():
+#         player_count = get_player_count()
+#         game = ScrabbleGame(player_count)
+#         while game.is_playing():
+#             show_board(game.get_board())
+#             show_player(*game.get_current_player())
+#             word, coords, orientation =get_inputs()
+#             try:
+#                 game.play(word, coords, orientation)
+#             except Exception as e:
+#                  print(e)